% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pdiff_ind_contrast.R
\name{estimate_pdiff_ind_contrast}
\alias{estimate_pdiff_ind_contrast}
\title{Estimate the magnitude of difference for an independent groups contrast.}
\usage{
estimate_pdiff_ind_contrast(
  data = NULL,
  outcome_variable = NULL,
  grouping_variable = NULL,
  events = NULL,
  ns = NULL,
  contrast = NULL,
  case_level = 1,
  outcome_variable_levels = NULL,
  grouping_variable_levels = NULL,
  grouping_variable_name = "My grouping variable",
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{grouping_variable}{For raw data - The column name of the grouping
variable, or a vector of group names}

\item{events}{For summary data - A vector of 2 or more event counts,
each an integer >= 0}

\item{ns}{For summary data - A vector of sample sizes, same length as counts,
each an integer >= correspond event count}

\item{contrast}{A vector of group weights}

\item{case_level}{For raw data, an optional numeric or character label
that specifies which level of the outcome variable to use for the
proportion.}

\item{outcome_variable_levels}{For summary data - An optional vector of
group labels, same length as counts}

\item{grouping_variable_levels}{For summary data - An optional vector of
group labels, same length as means}

\item{grouping_variable_name}{Optional friendly name for the grouping
variable.  Defaults to 'My grouping variable' or the grouping variable
column name if a data.frame is passed.}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{assume_equal_variance}{Defaults to FALSE}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_mdiff_ind_contrast} returns the point estimate and
confidence interval for the proportion difference in a linear contrast.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

estimate_pdiff_ind_contrast(
)

}
