% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pdiff_paired.R
\name{estimate_pdiff_paired}
\alias{estimate_pdiff_paired}
\title{Estimate magnitude of difference between two related measures.}
\usage{
estimate_pdiff_paired(
  data = NULL,
  comparison_measure = NULL,
  reference_measure = NULL,
  cases_consistent = NULL,
  cases_inconsistent = NULL,
  not_cases_consistent = NULL,
  not_cases_inconsistent = NULL,
  case_level = NULL,
  grouping_variable_levels = c("Comparison measure", "Reference measure"),
  conf_level = 0.95
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{comparison_measure}{For raw data - The comparison measure, a factor.  Can
be the column name of a data frame of a vector.}

\item{reference_measure}{For raw data - The reference measure, a factor.  Can
be the column name of a data frame of a vector.}

\item{cases_consistent}{Cases}

\item{cases_inconsistent}{Cases}

\item{grouping_variable_levels}{For summary data - An optional vector of
2 group labels.  Defaults to the names of the comparison and reference
measures, if passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{case_label}{Cases}
}
\value{
Returns object of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_pdiff_paired} returns effect sizes for comparing two
related nominal measures.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_mdiff_paired(
 iris,
 Sepal.Length,
 Petal.Length
)

}
