% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovidescribe.h.R
\name{jamovidescribe}
\alias{jamovidescribe}
\title{Describe}
\usage{
jamovidescribe(data, outcome_variable, show_details = FALSE)
}
\arguments{
\item{data}{.}

\item{outcome_variable}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$describe_plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Describe
}
