% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimdiffindcontrast.h.R
\name{jamovimdiffindcontrast}
\alias{jamovimdiffindcontrast}
\title{Mdiff - Independent Groups Contrast}
\usage{
jamovimdiffindcontrast(
  switch = "from_raw",
  data,
  outcome_variable,
  grouping_variable,
  means,
  sds,
  ns,
  grouping_variable_levels,
  outcome_variable_name = "Outcome variable",
  grouping_variable_name = "Grouping variable",
  comparison_labels = " ",
  reference_labels = " ",
  conf_level = 95,
  assume_equal_variance = TRUE,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{grouping_variable}{.}

\item{means}{.}

\item{sds}{.}

\item{ns}{.}

\item{grouping_variable_levels}{.}

\item{outcome_variable_name}{.}

\item{grouping_variable_name}{.}

\item{comparison_labels}{.}

\item{reference_labels}{.}

\item{conf_level}{.}

\item{assume_equal_variance}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_smd} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_difference} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Mdiff - Independent Groups Contrast
}
