% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimdiffpaired.h.R
\name{jamovimdiffpaired}
\alias{jamovimdiffpaired}
\title{Paired}
\usage{
jamovimdiffpaired(
  switch = "from_raw",
  data,
  comparison_measure,
  reference_measure,
  comparison_mean = " ",
  comparison_sd = " ",
  reference_mean = " ",
  reference_sd = " ",
  n = " ",
  correlation = " ",
  comparison_measure_name = "Comparison measure",
  reference_measure_name = "Reference measure",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{comparison_measure}{.}

\item{reference_measure}{.}

\item{comparison_mean}{.}

\item{comparison_sd}{.}

\item{reference_mean}{.}

\item{reference_sd}{.}

\item{n}{.}

\item{correlation}{.}

\item{comparison_measure_name}{.}

\item{reference_measure_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_r} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_smd} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_ratio} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_ratio} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Paired
}
