% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimdifftwo.h.R
\name{jamovimdifftwo}
\alias{jamovimdifftwo}
\title{Two groups}
\usage{
jamovimdifftwo(
  switch = "from_raw",
  data,
  outcome_variable,
  grouping_variable,
  comparison_mean = " ",
  comparison_sd = " ",
  comparison_n = " ",
  reference_mean = " ",
  reference_sd = " ",
  reference_n = " ",
  comparison_level_name = "Comparison level",
  reference_level_name = "Reference level",
  outcome_variable_name = "Outcome variable",
  grouping_variable_name = "Grouping variable",
  conf_level = 95,
  assume_equal_variance = TRUE,
  switch_comparison_order = FALSE,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{grouping_variable}{.}

\item{comparison_mean}{.}

\item{comparison_sd}{.}

\item{comparison_n}{.}

\item{reference_mean}{.}

\item{reference_sd}{.}

\item{reference_n}{.}

\item{comparison_level_name}{.}

\item{reference_level_name}{.}

\item{outcome_variable_name}{.}

\item{grouping_variable_name}{.}

\item{conf_level}{.}

\item{assume_equal_variance}{.}

\item{switch_comparison_order}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_smd} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_ratio} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_ratio} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Two groups
}
