% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimetamdiff.h.R
\name{jamovimetamdiff}
\alias{jamovimetamdiff}
\title{Mdiff - Meta-analysis}
\usage{
jamovimetamdiff(
  data,
  comparison_means,
  comparison_sds,
  comparison_ns,
  reference_means,
  reference_sds,
  reference_ns,
  labels,
  moderator,
  effect_label = "My effect",
  reported_effect_size = "mean_difference",
  conf_level = 95,
  random_effects = TRUE,
  show_details = FALSE
)
}
\arguments{
\item{data}{.}

\item{comparison_means}{.}

\item{comparison_sds}{.}

\item{comparison_ns}{.}

\item{reference_means}{.}

\item{reference_sds}{.}

\item{reference_ns}{.}

\item{labels}{.}

\item{moderator}{.}

\item{effect_label}{.}

\item{reported_effect_size}{.}

\item{conf_level}{.}

\item{random_effects}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a html \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$raw_data} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta_difference} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raw_data$asDF}

\code{as.data.frame(results$raw_data)}
}
\description{
Mdiff - Meta-analysis
}
