% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovipdiffindcontrast.h.R
\name{jamovipdiffindcontrast}
\alias{jamovipdiffindcontrast}
\title{Pdiff - Independent Groups Contrast}
\usage{
jamovipdiffindcontrast(
  data,
  outcome_variable,
  grouping_variable,
  means,
  sds,
  ns,
  grouping_variable_labels,
  outcome_variable_name = "Outcome variable",
  grouping_variable_name = "Grouping variable",
  comparison_levels,
  reference_levels,
  conf_level = 95,
  assume_equal_variance = TRUE
)
}
\arguments{
\item{data}{.}

\item{outcome_variable}{.}

\item{grouping_variable}{.}

\item{means}{.}

\item{sds}{.}

\item{ns}{.}

\item{grouping_variable_labels}{.}

\item{outcome_variable_name}{.}

\item{grouping_variable_name}{.}

\item{comparison_levels}{.}

\item{reference_levels}{.}

\item{conf_level}{.}

\item{assume_equal_variance}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Pdiff - Independent Groups Contrast
}
