% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovipdiffpaired.h.R
\name{jamovipdiffpaired}
\alias{jamovipdiffpaired}
\title{Paired}
\usage{
jamovipdiffpaired(
  switch = "from_raw",
  data,
  reference_measure,
  comparison_measure,
  cases_consistent = " ",
  cases_inconsistent = " ",
  not_cases_consistent = " ",
  not_cases_inconsistent = " ",
  case_label = "Sick",
  not_case_label = "Well",
  comparison_measure_name = "Pretest",
  reference_measure_name = "Posttest",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{reference_measure}{.}

\item{comparison_measure}{.}

\item{cases_consistent}{.}

\item{cases_inconsistent}{.}

\item{not_cases_consistent}{.}

\item{not_cases_inconsistent}{.}

\item{case_label}{.}

\item{not_case_label}{.}

\item{comparison_measure_name}{.}

\item{reference_measure_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_proportion_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_phi} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Paired
}
