% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovipdifftwo.h.R
\name{jamovipdifftwo}
\alias{jamovipdifftwo}
\title{Two Groups}
\usage{
jamovipdifftwo(
  switch = "from_raw",
  data,
  outcome_variable,
  grouping_variable,
  comparison_cases = " ",
  comparison_n = " ",
  reference_cases = " ",
  reference_n = " ",
  case_label = "Sick",
  not_case_label = "Well",
  grouping_variable_level1 = "Treated",
  grouping_variable_level2 = "Control",
  outcome_variable_name = "Outcome variable",
  grouping_variable_name = "Grouping variable",
  count_NA = FALSE,
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{grouping_variable}{.}

\item{comparison_cases}{.}

\item{comparison_n}{.}

\item{reference_cases}{.}

\item{reference_n}{.}

\item{case_label}{.}

\item{not_case_label}{.}

\item{grouping_variable_level1}{.}

\item{grouping_variable_level2}{.}

\item{outcome_variable_name}{.}

\item{grouping_variable_name}{.}

\item{count_NA}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_proportion_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_odds_ratio} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Two Groups
}
