% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamoviproportion.h.R
\name{jamoviproportion}
\alias{jamoviproportion}
\title{Single Group}
\usage{
jamoviproportion(
  switch = "from_raw",
  data,
  outcome_variable,
  cases = " ",
  observations = " ",
  case_label = "Affected",
  not_case_label = "Not Affected",
  outcome_variable_name = "Outcome variable",
  count_NA = FALSE,
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{cases}{.}

\item{observations}{.}

\item{case_label}{.}

\item{not_case_label}{.}

\item{outcome_variable_name}{.}

\item{count_NA}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Single Group
}
