% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovirdiffpaired.h.R
\name{jamovirdiffpaired}
\alias{jamovirdiffpaired}
\title{Paired}
\usage{
jamovirdiffpaired(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Paired
}
