% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovirdifftwo.h.R
\name{jamovirdifftwo}
\alias{jamovirdifftwo}
\title{Two groups}
\usage{
jamovirdifftwo(
  switch = "from_raw",
  data,
  x,
  y,
  grouping_variable,
  comparison_r = " ",
  comparison_n = " ",
  reference_r = " ",
  reference_n = " ",
  x_variable_name = "X variable",
  y_variable_name = "Y variable",
  comparison_level_name = "Comparison level",
  reference_level_name = "Reference level",
  grouping_variable_name = "Grouping variable",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{x}{.}

\item{y}{.}

\item{grouping_variable}{.}

\item{comparison_r}{.}

\item{comparison_n}{.}

\item{reference_r}{.}

\item{reference_n}{.}

\item{x_variable_name}{.}

\item{y_variable_name}{.}

\item{comparison_level_name}{.}

\item{reference_level_name}{.}

\item{grouping_variable_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_r} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Two groups
}
