% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovirmetamean.h.R
\name{jamovirmetamean}
\alias{jamovirmetamean}
\title{Means}
\usage{
jamovirmetamean(
  data,
  means,
  sds,
  ns,
  labels,
  moderator,
  reference_mean = "0",
  effect_label = "My effect",
  reported_effect_size = "mean_difference",
  conf_level = 95,
  random_effects = TRUE,
  show_details = FALSE
)
}
\arguments{
\item{data}{.}

\item{means}{.}

\item{sds}{.}

\item{ns}{.}

\item{labels}{.}

\item{moderator}{.}

\item{reference_mean}{.}

\item{effect_label}{.}

\item{reported_effect_size}{.}

\item{conf_level}{.}

\item{random_effects}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a html \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$raw_data} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta_difference} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raw_data$asDF}

\code{as.data.frame(results$raw_data)}
}
\description{
Means
}
