% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovirmetar.h.R
\name{jamovirmetar}
\alias{jamovirmetar}
\title{Linear Correlation - Meta-analysis}
\usage{
jamovirmetar(
  data,
  rs,
  ns,
  labels,
  moderator,
  effect_label = "My effect",
  conf_level = 95,
  random_effects = TRUE,
  show_details = FALSE
)
}
\arguments{
\item{data}{.}

\item{rs}{.}

\item{ns}{.}

\item{labels}{.}

\item{moderator}{.}

\item{effect_label}{.}

\item{conf_level}{.}

\item{random_effects}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a html \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$raw_data} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta_difference} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raw_data$asDF}

\code{as.data.frame(results$raw_data)}
}
\description{
Linear Correlation - Meta-analysis
}
