% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_any.R
\name{meta_any}
\alias{meta_any}
\title{Estimate any meta effect}
\usage{
meta_any(
  data,
  moderator = FALSE,
  effect_size_name = "MDiff",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble with columns
\itemize{
\item label, LL, UL, yi, vi, moderator and any other columns relevant to that meta-analysis
}}

\item{moderator}{logical if moderator has been passed}

\item{effect_size_name}{defaults to MDiff}

\item{random_effects}{logical}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_any} returns
}
