% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_mdiff_two.R
\name{meta_mdiff_two}
\alias{meta_mdiff_two}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_mdiff_two(
  data,
  comparison_means,
  comparison_sds,
  comparison_ns,
  reference_means,
  reference_sds,
  reference_ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  report_smd = FALSE,
  random_effects = TRUE,
  assume_equal_variance = FALSE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{comparison_means}{comparison}

\item{comparison_sds}{comparison}

\item{reference_means}{reference}

\item{reference_sds}{reference}

\item{reference_ns}{reference}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{el}

\item{report_smd}{smd}

\item{random_effects}{re}

\item{assume_equal_variance}{aev}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{comparison_n}{comparison}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_mdiff_two} returns
}
