% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_smd_two.R
\name{meta_smd_two}
\alias{meta_smd_two}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_smd_two(
  data,
  smds,
  comparison_ns,
  reference_ns,
  moderator = NULL,
  contrast = NULL,
  labels = NULL,
  effect_label = "My effect",
  random_effects = TRUE,
  assume_equal_variance = FALSE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{smds}{smd}

\item{reference_ns}{reference}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{labels}{labels}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{comparison_n}{comparison}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_smd_two} returns
}
