% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_nominal.R
\name{overview_nominal}
\alias{overview_nominal}
\title{Calculates descriptive statistics for a numerical variable}
\usage{
overview_nominal(
  data = NULL,
  outcome_variable = NULL,
  countss = NULL,
  ns = NULL,
  outcome_variable_levels = NULL,
  outcome_variable_name = "My Outcome Variable",
  conf_level = 0.95,
  assume_equal_variance = FALSE
)
}
\arguments{
\item{data}{\itemize{
\item for raw data, a data frame or tibble
}}

\item{outcome_variable}{\itemize{
\item for raw data, either a vector containing factor
data or the name of a data-frame column containing a factor
}}

\item{ns}{For summary data - A vector of sample sizes, same length as counts}

\item{outcome_variable_levels}{For summary data - An optional vector of
group labels, same length as counts.  If not passed, auto-generated.}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My Outcome Variable'.  Ignored if a data-frame is passed,
this argument is ignored.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{counts}{For summary data - A vector of 1 or more counts, integers,0}
}
\value{
Returns a table of descriptive statistics
}
\description{
This function caluclated basic descriptive statistics for a categorical/
nominal variable. Inputs can be summary data, vectors, or a data frame.
}
\examples{
# From Summary data --------------------------------------

overview_nominal(
  counts = c(10, 15, 100),
  ns = c(100, 100, 100),
  outcome_variable_levels = c(
     "First_Years", "Second_Years", "Third_Years"
  ),
  outcome_variable_name = "\% time near target",
  assume_equal_variance = TRUE
)


}
