% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovicorrelation.h.R
\name{jamovicorrelation}
\alias{jamovicorrelation}
\title{Correlation}
\usage{
jamovicorrelation(
  switch = "from_raw",
  data,
  x,
  y,
  r = " ",
  n = " ",
  x_variable_name = "x variable",
  y_variable_name = "y variable",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{x}{.}

\item{y}{.}

\item{r}{.}

\item{n}{.}

\item{x_variable_name}{.}

\item{y_variable_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_r} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Correlation
}
