% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimagnitude.h.R
\name{jamovimagnitude}
\alias{jamovimagnitude}
\title{Magnitude/Quantity}
\usage{
jamovimagnitude(
  switch = "from_raw",
  data,
  outcome_variable,
  mean = " ",
  sd = " ",
  n = " ",
  outcome_variable_name = "Outcome variable",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{mean}{.}

\item{sd}{.}

\item{n}{.}

\item{outcome_variable_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Magnitude/Quantity
}
