% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimdiffone.h.R
\name{jamovimdiffone}
\alias{jamovimdiffone}
\title{Mdiff - Single Group}
\usage{
jamovimdiffone(
  switch = "from_raw",
  data,
  outcome_variable,
  comparison_mean = " ",
  comparison_sd = " ",
  comparison_n = " ",
  reference_mean = " ",
  outcome_variable_name = "Outcome variable",
  conf_level = 95,
  show_details = FALSE
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{comparison_mean}{.}

\item{comparison_sd}{.}

\item{comparison_n}{.}

\item{reference_mean}{.}

\item{outcome_variable_name}{.}

\item{conf_level}{.}

\item{show_details}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_smd} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Mdiff - Single Group
}
